[KRROOD] Adressed feedback
